+++ src/uncms/media/fields.py
@@ -3,7 +3,7 @@ from django.contrib.admin.widgets import ForeignKeyRawIdWidget
 from django.db import models
 
 from uncms.conf import defaults
-from uncms.media.filetypes import IMAGE_FILENAME_REGEX
+from uncms.media.filetypes import IMAGE_FILE_EXTENSIONS
 from uncms.media.widgets import ImageThumbnailWidget
 
 
@@ -21,32 +21,40 @@ class FileRefField(models.ForeignKey):
         return super().formfield(**kwargs)
 
 
-class ImageRefField(FileRefField):
-    '''A foreign key to a File, constrained to only select image files.'''
+class RestrictedFileRefField(FileRefField):
+    '''
+    A FileRefField that only allows files of certain extensions.
+    '''
+    allowed_extensions = []
 
     def __init__(self, **kwargs):
-        # we have to use the regex here because Q objects don't work in
-        # related popups in the admin
-        kwargs['limit_choices_to'] = {
-            'file__iregex': IMAGE_FILENAME_REGEX,
-        }
+        assert self.allowed_extensions, 'using RestrictedFileRefField without allowed_extensions does not make sense'
 
+        kwargs['limit_choices_to'] = {
+            'file__iregex': ''.join([
+                r'\.',
+                '(',
+                "|".join(self.allowed_extensions),
+                ')',
+                '$',
+            ])
+        }
         super().__init__(**kwargs)
+
+
+class ImageRefField(RestrictedFileRefField):
+    '''
+    A foreign key to a File, constrained to only select image files.
+    '''
+    allowed_extensions = IMAGE_FILE_EXTENSIONS
 
     def formfield(self, **kwargs):
         kwargs.setdefault('widget', ImageThumbnailWidget(self.remote_field, admin.site))
         return super().formfield(**kwargs)
 
 
-VIDEO_FILTER = {
-    'file__iregex': r'\.(mp4|m4v)$'
-}
-
-
-class VideoFileRefField(FileRefField):
-
-    '''A foreign key to a File, constrained to only select video files.'''
-
-    def __init__(self, **kwargs):
-        kwargs['limit_choices_to'] = VIDEO_FILTER
-        super().__init__(**kwargs)
+class VideoFileRefField(RestrictedFileRefField):
+    '''
+    A foreign key to a File, constrained to only select video files.
+    '''
+    allowed_extensions = ['mp4', 'm4v']